cleaner __init__ for oracle.
